+++ src/generators/files/server-ts.ts
@@ -1,10 +1,13 @@
 import * as models from "../../models/index.js";
+import { toCamel, toPascal } from "../../utils/index.js";
 import { itt } from "../../utils/iterable-text-template.js";
-import { generateIsAuthenticationCode } from "../functions/index.js";
+import { generateIsAuthenticationFunctionBody } from "../bodies/index.js";
 import {
-  generateAuthenticationTypesCode,
-  generateOperationsTypeCode,
-  generateServerAuthenticationTypeCode,
+  generateOperationAuthenticationType,
+  generateOperationHandlerType,
+  generateOperationIncomingRequestType,
+  generateOperationOutgoingResponseType,
+  generateServerAuthenticationType,
   generateServerClass,
 } from "../types/index.js";
 
@@ -15,9 +18,49 @@ export function* generateServerTsCode(apiModel: models.Api) {
     import * as lib from "@oa42/oa42-lib";
   `;
 
-  yield* generateAuthenticationTypesCode(apiModel);
-  yield* generateServerAuthenticationTypeCode(apiModel);
-  yield* generateOperationsTypeCode(apiModel);
+  yield* generateServerAuthenticationType(apiModel);
   yield* generateServerClass(apiModel);
-  yield* generateIsAuthenticationCode(apiModel);
+
+  for (const authenticationModel of apiModel.authentication) {
+    const handlerTypeName = toPascal(
+      authenticationModel.name,
+      "authentication",
+      "handler",
+    );
+
+    yield itt`
+      export type ${handlerTypeName}<A extends ServerAuthentication> =
+        (credential: string) => A[${JSON.stringify(
+          toCamel(authenticationModel.name),
+        )}];
+    `;
+  }
+
+  for (const pathModel of apiModel.paths) {
+    for (const operationModel of pathModel.operations) {
+      const isAuthenticationFunctionName = toCamel(
+        "is",
+        operationModel.name,
+        "authentication",
+      );
+      const authenticationTypeName = toPascal(
+        operationModel.name,
+        "authentication",
+      );
+
+      yield itt`
+        export function ${isAuthenticationFunctionName}<A extends ServerAuthentication>(
+          authentication: Partial<${authenticationTypeName}<A>>,
+        ): authentication is ${authenticationTypeName}<A> {
+          ${generateIsAuthenticationFunctionBody(pathModel, operationModel)}
+        }
+      `;
+
+      yield* generateOperationAuthenticationType(operationModel);
+      yield* generateOperationHandlerType(operationModel);
+
+      yield generateOperationIncomingRequestType(apiModel, operationModel);
+      yield generateOperationOutgoingResponseType(apiModel, operationModel);
+    }
+  }
 }
