--- src/generators/functions/client-operations.ts
@@ -1,29 +1,0 @@
-import * as models from "../../models/index.js";
-import { toCamel } from "../../utils/index.js";
-import { itt } from "../../utils/iterable-text-template.js";
-import { generateClientOperationFunctionBody } from "../bodies/index.js";
-
-export function* generateClientOperationsCode(apiModel: models.Api) {
-  yield* generateAllFunctions(apiModel);
-}
-
-function* generateAllFunctions(apiModel: models.Api) {
-  for (const pathModel of apiModel.paths) {
-    for (const operationModel of pathModel.operations) {
-      yield* generateFunction(pathModel, operationModel);
-    }
-  }
-}
-
-function* generateFunction(
-  pathModel: models.Path,
-  operationModel: models.Operation,
-) {
-  const name = toCamel(operationModel.name);
-
-  yield itt`
-    export function ${name}(){
-      ${generateClientOperationFunctionBody(pathModel, operationModel)}
-    }
-  `;
-}
