--- src/generators/functions/index.ts
@@ -1,3 +1,0 @@
-export * from "./client-operations.js";
-export * from "./is-authentication.js";
-export * from "./is-parameters.js";
