+++ src/generators/types/index.ts
@@ -1,5 +1,8 @@
 export * from "./authentications.js";
-export * from "./operations.js";
+export * from "./operation-authentication.js";
+export * from "./operation-handler.js";
+export * from "./operation-incoming-request.js";
+export * from "./operation-outgoing-response.js";
 export * from "./parameters.js";
 export * from "./server-authentication.js";
 export * from "./server.js";
